Fix #148562: Overflow when uploading pointclouds to GPU
An overflow can happen when a huge amount of points are drawn. Each
point is triangulated and the index buffer can only store 4GB at max.

During upload it is now detected if the index buffer could safely be
stored on the GPU. If this isn't the case it will draw no points for the
pointcloud.

Many improvements can be made in this area to provide better feedback to
the user. This fix only fixes the write out of bound. Suggestions are:

- Display a message in the GN/node
- Separate into multiple vertex/index buffers to relax the limitation
- Replace index buffer with a vertex shader