Fix: Some animation editors should show the footer by default
Update the UI versioning code that adds the new footer to the Dope Sheet
and Video Sequencer, to also show the footer by default. This now
matches their defaults when new editors are created.

-------

TODO: see what Aras answered in the Sequencer chat about this, because
the defaults when a new editor are created are inconsistent with the
Video Editor workspace. The above commit message already assumes this
inconsistency has been fixed by showing the footer by default.