Fix: Do not show the footer region in the Driver Editor
When the Graph Editor is in Driver mode, the X-axis no longer represents
time. It should therefore also not show the footer with the playback
controls. This was already taken into account by the drawing of those
controls, but the footer was still there (but empty).

A poll function now prevents it from showing up, and the View menu no
longer shows option to show/hide the Playback Controls.