Python: Stop HTTP Background Downloader process when Blender dies
When Blender dies in a non-clean way (which prevents it from shutting
down background processes gracefully), this is now detected by the HTTP
Background Downloader process, and makes it shut down.

This is done by catching the `EOFError` and `BrokenPipeError` thrown
when receiving resp. sending messages from/to the main process.