Fix: Don't Maximize Area When Single Editor in Window
"Maximize Area" (Ctrl-space) makes an area take up the entire window,
normally showing a "Back to Previous" button on the Top Bar header.
This PR makes it so that we don't enter this mode if there is only
one editor in the window as the area is already maximized. This does
not affect "Focus" mode. This change does not make any visible change
to 5.0 but does fix a one-pixel shift in 4.5 (#144740) so we can
possibly backport this eventually.