Refactor: simplify code in ED_region_is_overlap()
Reduce cognitive complexity by returning early (if the
`USER_REGION_OVERLAP` flag is set, just return `false`), use a `switch`
instead of `if`/`else if`/`else` chain, and just `return condition`
instead of `if (condition) return true`.

The result of this is that you have to read significantly less code to
understand what is being returned at any point in the function.

No functional changes.