Assets: periodically download remote asset listings
Re-download remote asset library listings every week. This is checked,
just like done with the Extensions platform, on Blender start-up.

When a remote listing is downloaded completely, i.e. the last page has
been downloaded, validated, and parsed, the top metadata file is
touched. Its timestamp serves as an indicator of the last succesful
download of the listing.