Fix #137111: Solo scrubbing can not be initiated on strip handle
Add `fallthrough_on_strip_handles` operator property to facilitate this.
The property is needed, because with RCS the handle tweaking has priority over
scrubbing. However in LCS scrubbing is initiated with RMB, so this could be
allowed, but this has to be defined explicitly. Well I could check event, but
having property for this is better.