Fix #146125: Some Simplified Chinese tooltips have two periods
In Blender, a tooltip can currently end with a punctuation sign
from (".", "!", "?"). If it doesn't, a period will be added.

Some languages use other sentence-end punctuation symbols, and those
will not be detected so a period will be added anyway. In Chinese,
"． " and "。" are affected.

This commit makes those signs be detected to avoid having two
periods in Chinese translations.