I18n: Make single tooltip translatable in brush stroke method
This tooltip from an enum item was not translatable because it is part
of a dynamic enum item, so not extracted. The other items from the
same enum were however, because they are exact duplicates to another
enum, sculpt_stroke_method.

This commit simply fixes the typo so the tooltip is aligned with
sculpt_stroke_method_items.