Fix #148736: LineArt: Correct vertex weight transfer implementation
For some reason during migration to Grease Pencil v3 the vertex weight
transfer part of line art was never implemented correctly (likely was a
temporary solution to get the feature working), and the name-prefix
based filter was never migrated correctly. This fix included the
correction for such vertex group filtering logic that is consistent with
what we have prior to GPv3/Blender v4.3.