Fix #148478: renaming bone can rename channels in all action slots
If an action is assigned to an action constraint on a bone, renaming
that bone would rename matching channels in all slots on the action,
not just the assigned slot.

The issue was simply that the code that did the f-curve renaming didn't
account for slots at all. This fixes the issue by adding an appropriate
`slot_handle` parameter to the relevant function.