WIP: Fix #148822: Crash undoing with sequencer preview open
There's some code in the memfile undo system that tries to restore
dependency graphs after loading an undo step.

This apparently also restored the depsgraph built for the sequencer
preview which causes a crash later on.

This fixes the issue by discarding the depsgraph after an undo step
has been decoded.