Fix #148791: Make link operator fails for simple cases
The Make Link operator fails in simple cases like the Convolve node,
where a "high priority" socket like the menu socket is available,
which makes the code skip the first input due to an error in the logic
of the function that finds the best input.

The logic in the function is too complex for what it tries to do, so a
fix in the current state is not worth it. So we rewrite the function in
an easier form.