Fix #146700: Preserve better Materials orders per object.
The limitations of FBX format when it comes to Face material indices
(they map to the order in which materials are connected to the object
using the mesh) makes it impossible to guarantee a correct, preserved
order of Blender material indices at export in the geometry.

However, by iterating over the objects and then their materials will
help giving better, more consistent results, and address issues like the
one reported in #146700.

(The previous process was iterating over all gathered materials, and
then their user objects, which could lead to re-ordering of the
materials in some objects under certain circumstances).