Small optimizations
Most of these resulted in only very tiny or immesurable speed increases.
This is likely because the compiler does the same optimizations for you.

SSE2 implementation for 4 components, timing maybe improved about 1%.
This pretty much shows that the gcc compiler is able to optimise the float4
math to use the same instructions. I kept it in the code anyway in case it
helps other systems.

Other attempts at optimizations, such as initializing sum to the first
value rather than adding the first value to a zero sum, resulted in poorer
performance. Best guess is this is due to more complex code making the
compiler not optimize as well, and/or instruction cache or branch prediction
not working as well.