Fix #146128: Relocate Output Properties Warning
Properties / Output / Output / Encoding / Video can show two different
warnings, but they are both shown below "Color Space" and indented so
they can be easily cut off. This PR just moves the two warnings to the
top of the section, left-aligned. This seems to be a better way to
indicate that the current settings need to change.