UI: Consistent Area Movement Snapping
Holding Ctrl while dragging area edges will cause snapping at the
extremes and on twelfths. Often unnoticed is that these divisions are
along the unobstructed travel distance. So if you can drag without
being stopped by another area then 50% will be at the screen center.
But if another area is positioned in such a way to stop you from
dragging then the snapping stops are distributed along the available
distance, and therefore 50% will be a different place. This PR makes
the snapping stops always relative to the screen area. So the 50%
position is always in the center, for example.