Fix #148695: Value slider dragging lag due to missing redraw
While this issue was first reported as being caused by 8c9025d9aa,
which made dragging operation use a different event method
for getting the mouse location on macOS (mouseLocation vs.
mouseLocationOutsideOfEventStream), this commit only exposed a
much older issue with value sliders missing to redraw when mouse
events were sent at high rates.

Investigating this in depth through the interface button handlers, the
culprit lies in the `ui_but_dragedit_update_mval()` function, which is
used to determined if a mouse event should be taken into account when
dragging a numeric button/slider, to then apply the value and redraw the
button if that's the case.

At the start of this function, a check is made to check if the passed in
mx value isn't the same as the saved `draglastx` value. The problem
being that for most of the drag events going through this function,
draglastx is being constantly updated by other branches, and ends up
already matching with the mx value if multiple mouse events with the
same X values end up being sent at a high rate, bypassing the
update/redraw.

Interestingly enough, this function is called in two places, for
draggable number sliders and draggable number buttons. However, as the
original issue report, only sliders are affected. With the reason being
that in the case of number buttons, a screen_mx screen coord value is
passed to the function in `ui_do_but_NUM`, while `draglastx` is still
being updated with the "normal" mx value, bypassing the check.

Fixed by removing this check, a lot more improvements could also be made
cleaning up, modernizing and abstracting shared behavior throughout the
interface button handlers code, but this can be done separately. While
this was marked as a macOS bug, in practice this could also fix issues
with number slider missing redraws on other OSes too.