Fix (unreported) remapping of owner ID pointer in embedded IDs.
Just like pinters from owner IDs to their embedded IDs should never be
remapped, the loopback one should not either. These relations between
IDs and their embedded data are the responsibility of the owner ID's
internal data management, _not_ of the generic ID management code.

Not sure how that managed to fly under the radar for so long, as this
was a fairly bad bug...

Discovered while investigating #148554.