Fix #148914: GPU: Curve out of bound read
Do not draw the last restart segment for
ribbon and line strip topology.

Following the documentation from #142969
the point attribute buffers (like position+radius)
do not contain the restart indices. As such
the restart segment always load past the current
curve data end. This is an issue for the last curve
since there is not data after the last element, creating
an out of bound read.

Removing the useless last restart segment avoid loading
the out of bound data. Note that this isn't an issue with
cylinder primitive type as they do not have restart segments.