Fix #148980: macOS: Continuous grab wrapping fails when no custom bounds are provided
Similarly to 6d87fcdaf5, this issue was also indirectly caused by
8c9025d9aa exposing an older issues where using window / non-custom
bounds for continuous cursor grab would fail to wrap, causing the cursor
to get stuck at the edge of the screen.

Before 8c9025d9aa this issue would only be visible for non-hidden cursor
grabbing operations with no custom bounds, which weren't used. Now that
hidden cursor grabbing operations are also using wrapping, sometimes
without custom bounds, as is the case of number sliders, the issue
appears.

Internally, the issue is caused by the Cocoa wrapping assuming grab
bound coordinates to use a bottom-left origin, and converting them back
to top left. However, when no custom grab bounds are provided, window
bounds are used as a fallback, which are already using a top-left
origin. The origin conversion is then wrongly applied, flipping it to
bottom-left and preventing proper wrapping.

Fixed by adding a simple condition to check for bottom-left coordinates
before applying the conversion. Ideally, I'd prefer for this conversion
to not be done this late, and for all set grab bounds to be normalized
to the right origin ahead of their use. Digging a bit more, the real
*core* of the issue lies in the fact that for custom bounds, the
`wm_cursor_position_to_ghost_screen_coords()` function is called in
`WM_cursor_grab_enable`, causing the initial coordinate flip on macOS.

The proper fix would thus be to remove this potentially superflous call,
however doing so would make for a more complex change that also affect
other backends. So this simple fix will do for now, especially close to
the 5.0 release.