Fix #147159: Crash selecting brush asset from current file after renaming
This sanitizes asset system code to better deal with renaming asset
data-blocks, preventing the crash. It's not a perfect solution, since the brush
won't be highlighted as active anymore after renaming. This is because the
highlighting is based on `AssetWeakReference`, which is known to break on
actions like renaming (hence the "weak" in the name). It still holds the old
name. I think we can make the highlighting use an asset reference from the
actual active brush to fix this. For now focus on preventing the crash.