Fix #110618: Reorder Walk & Fly Keymap Items
The keymap entries for Walk Navigation contains many that have separate
"on" and "off" keys. There is one to start to move forward and another
to stop forward movement. Unfortunately it is easy to change one and
not the other because they are very separated in the keymap editor.
This PR just sorts these items so that they are next to each other.
Although fly navigation does not also have on/off this PR sorts its
items so that these are in a similar order.