Fix: Nodes: Links not properly preserved when swapping between two node groups ---
There was an early return in the swap operator code if a node is being swapped with its own type.

This didn't account for node groups and assets however, which all use the same Group node. Since different node groups can have wildly varying sockets & properties, it makes sense to not include them in the early return.

This fixes the problem of various links not being properly preserved when swapping between two node groups.