Fix #144727: Cycles adaptive subdivision with displacement + bump wrong
Need to do bump mapping on the undisplaced normal. Previous the regular
smooth normal was already undisplaced, but this is no longer the case.

We would be able to save memory if that behavior was restored, but it is
complicated and not something to fix this close to the release.