Fix: 'Clear and Keep Transforms' enum value warning
When 'Clear and Keep Transforms' is used to remove parentage of an 
object and the object has a parent_type other than 'Object', any 
interaction that forces reevaluation (selecting the object, mouseover 
the 'parent type' field, etc) triggers the error:

03:39:19.562  bpy.rna          | WARNING current value '<value>' 
matches no enum in 'Object', '<obj.name>', 'parent_type'

with '<value>' == the parent's enum, regardless of when the 
parentage was removed. The only way to fix the warning constantly
 printing is to reparent the object to something else or manually set 
'Object' as the parent type. 

This fix just sets the parent type to the Object default when 'clear 
and keep transforms' is used, as is done for the regular 'clear parent'.

Warning began sometime between v4.4.1 and v4.5.3, still present in 5.1.0a