Fix: New node group gets socket default name instead of label
When a node group is created from a single node, it creates sockets
based on the node's declaration as well. It should use the socket's
label but instead used the default name.

For example, a group created from a `ShaderNodeMath` or
`FunctionNodeCombineColor` should have its socket names depend on the
mode / operation.

Noticed by Jacques Lucke in !149000.