Fix #149125: Preserve Pitch on retimed sounds broken when anything is animated
Similar case as was with equalizer modifier recently (#141084):
here the AnimateableTimeStretchPitchScale effect was being
constantly re-created every frame, as soon as anything in VSE
timeline has an animation.

Fix this by checking if Audaspace sequence handle for the sound
is already an instance of AnimateableTimeStretchPitchScale; if it
is then just use the already created one.