Fix #143187: UI: Previous tree view item stays selected after some operations
When the active item was changed externally (e.g. the operator to add a shape
key would activate the new shape key), there was code to de-activate and
de-select the previous active item. But the change was not synced to the
represented data (e.g. selection flag in shape keys data). This would cause the
item's state to be overridden as selected again (as it should).