Fix: Incorrect Icon Used for Collection in Asset Browser
We have icons that represent specific individual collections. like
Icon_Outliner_Collection for a default (uncolored) collection, and
Icon_Collection_color_x for ones with colors. For "collections" as a
general thing though we have icon_group. The incorrect one is used in
the filter list of the Asset Browser.