Vulkan: Reduce memory pressure when uploading textures
Memory pressure is a pain point of vulkan backend. There are extensions
that can help with this. One of them is VK_EXT_host_image_copy that
reduces the need of a staging buffer to upload data to an image. It
allows directly uploading from an application owned buffer and perform a
layout transition afterwards.

Current support contains NVIDIA, Intel and Qualcomm. AMD support is in
development. When using MESA it can be activated by setting the
`RADV_PERFTEST=hic` environment variable. It is currently disabled as it
doesn't improve on non linear surfaces. Blender uses non linear surfaces
so this is also the case for us.