Fix #149201: Cloth sim init writes to original mesh but is deform-only
The cloth modifier is making a temp copy of the input mesh for the purpose
of calculating rest positions and edge lengths. This was actually writing
to the original mesh data, creating a new positions array accidentally
through the implicit sharing process. This is a problem because the cloth
modifier is a `DeformOnly` modifier and changing mesh data during evaluation
can free the original mesh data which is still used by the modifier stack
evaluation.

In addition to using the correct mesh copy for computing rest positions,
the mesh is now also `const` throughout the modifier functions to avoid
future problems of this kind.