Fix #147539: Geometry Nodes Ctrl middle mouse up/down string crashes
Search button handling would call button applying more than once, but
after the first call the "search argument" was moved out of the button.

This moving of data and ownership is done to prevent dangling pointers,
since the button initially owning the data may be destructed before the
delayed button applying is run (delayed until all button handling is
done). But as noted eariler, this breaks applying the button multiple
times as part of the button handling, which is an otherwise sensible
thing to do.

In similar situations we duplicate memory (see
`uiBut.func_argN_copy_fn`), but introducing this duplication is quite
an intrusive change, and isn't a great solution anyway.

A memory-safe, more efficient and less intrusive solution is using
shared ownership, where multiple parts of the code can "own" the same
data, and it's only destructed with the last owner. (This approach
should be applied to other similar cases too, enabled by recent work
towards "modern" C++.)