Fix: Can not cancel waveform reading job
When there are lot of sound strips in VSE timeline, waveform rendering
could not be stopped. This causes problems mainly when .blend file was
opened by accident and different one can not be opened until all
waveforms are read, which can take a minute.

This was caused by quite weird code, where value of `stop` boolean
pointer was stored in local variable and cast into short. Pointer to
this local variable was passed to `AUD_readSound()`. Then value of
`stop` bool pointer was set to that local variable again.

Perhaps it was luck, but it seemed to me, that `AUD_readSound()` somehow
did interrupt its process when `stop` was changed to true, which it
shouldn't. But either way, the code spends 99.99% of time in that
function, so value of `stop` being overwritten from true to false is
pretty much certain.

The fix is to pass original pointer using `reinterpret_cast` and don't
modify it.

Finally, while the waveform reading job now can be cancelled, new job
will be created immediately. This is technicaly correct behavior, even
though it can cause other problems described in #144082