Fix #149288: DRW: Hair in linked collections are at the wrong location
This was caused because we use the evaluated object
world to object matrix to undo the psys transform.
But this matrix is not necessarily the one used
during the psys eval. Using the correct matrix
fixes the issue.

This `psys->imat` is also used by Cycles, so it
should be the correct one. Somehow, the draw engines
never used it and instead undo the dupli object
transform (see `ObjectRef::particles_matrix()`).