Compositor: Hide essentials from asset shelf meant for menu only
These essentials assets are meant to be exposed in the "Add" menu but
don't really fit into the asset shelf. They are just implemented as
assets because they can be created from other nodes, but still serve a
low level purpose. They cannot be represented visually well, so they
don't have an asset preview. For the same reason we don't have an asset
shelf for geometry nodes at all, and only display them in the menu.
Displaying them alongside other essentials assets that have a preview in
the asset shelf looked odd.