Update the AUTHORS file with the orgs from credits_git_gen.py
For the orgs that I did not find a contributor with a valid e-mail I
used <> (which is still a valid pattern, it just won't match anything).

---

Note I still need to bring PICO/ByteDance here, but I am waiting for Hogan reply about
which one to use (his email was bytedance though).