USD: Process point cloud and point instancer ID values
These Ids are required for correct in-camera motion blur so read and
write them as appropriate for both point clouds and point instancers.

The incoming USD Ids are 64-bit but Blender only supports 32-bit Ids so
the values will be narrowed. This will potentially produce different or
duplicate Ids depending on how far the value is outside the
[-2147483648, 2147483647] range.