Tools: Escape author HTML in git credits
Authors in credits_git_gen.py includes email addresses if the
`--use-email` flag is passed. These emails are between brackets. In
the final HTML they are considered as HTML tags because of these
brackets, which results in weird nested HTML and invisible addresses.

To avoid this, escape the author so that `<name@domain.tld>` becomes
`&lt;name@domain.tld&gt;`.