Assets: Make objects local when dropping, allows transforming
When dropping objects or collections into the 3D viewport, make them
local if the Packing import method is used, but keep any dependencies
(the mesh, assigned materials, etc) packed. This way the object can be
transformed, and the drag & drop snapping with the visual boundingbox
can work.

Also keep the collection packed if collection instancing is enabled.
Then the individual objects don't need to be transformed.

The introduction of the "Packing" import method removed to option to use
"Append (Reuse Data)". This means simply dragging an object to a 3D
Viewport wouldn't make it possible to transform the object. Neither
would the snapping grid and boundingbox show while dragging, as reported
in #149365.

Making objects and collections local as part of the dropping re-enables
that. Meshes, materials and other data-blocks used by the objects &
collections will remain packed, avoiding duplicates. Materials will
remain packed for now, they can be made local manually to do edits.