Assets: Drop material on object when object data is linked/packed
Includes !149472.

When dropping material on an object that has linked or packed object
data, make the material be assigned to the object, not the object data.

Without this, when dragging an object from the asset browser, and then a
material on it, the material would be non-editable because it's packed -
so far this is by design. But clicking the "Unpack" icon would not do
anything, because the material is linked to the object data, which is
non-editable too.

With this and !149472 applied, materials of dragged in objects can now
be replaced by drag & drop or edited more easily.