Fix: EEVEE: Broken Spherical Harmonic deringing
The Spherical Harmonic deringing code did not take
into account the case where one or more component
could be null. This would cause the normalization factor
to become 0 and destroy all directional components.

The fix is to clamp the constant term to a minimum
value (`1e-8` here) so that these null or quazi null
channels result in a normalization factor of 1.

The other issue was that the normalization factor
was computed by using `reduce_max` instead of `length`.
This caused different normalization factors as the
world lighting was rotating.

The last issue is just a simple improvement but we
now take into account the lambert convolution
weighting when searching for the maximas. This
allows to keep strong lighting sharper. However,
this can lead to less realistic lighting. So we
bias the weights a bit to avoid too much contrast
and retain a realistic look.