Fix #149443: Crash using HIP RT on recent GPU drivers
At the moment, HIP RT tries to load a HIP 6.X runtime, while the rest
of Cycles tries to load a HIP 7.X runtime, which then leads to
undefined behaviour and crashing.

To fix this issue, this commit prioritizes loading the HIP 6.X runtime
and falls back to the HIP 7.X runtime if 6.X doesn't exist.

This resolves crashing in currently avaliable GPU drivers, however the
issue will appear again once the GPU driver stops shipping with a HIP
6.X runtime.