Vulkan: Disable shader texture atomic test
Currently in main the shader texture atomic test fails. The root cause
is 0574c4b6f9606e131f022fe2575ada27378b7c90 that doesn't provide enough
information to the second dispatch to create the correct barriers.

The fix can take some time and is in the making. Until it has been fixed
the test will be disable to not intervene with the PRs of others.