Fix #147544: Crash during drawing while scene is switching
This crash was happening because the draw operator wasn't written
with the possibility in mind, that the context of the current scene
might change during modal execution.

Unfortunately this is the case for other operators as well.
This fix is meant to address the issue with as little impact on the
current state of the code as possible. While this will resolve the crash
that was reported, other scenarios might cause a crash as well, but
they might be less likely for users to run into for now.

The overall issue will have to be addressed outside of high severity
fixes during the next release cycle.