Fix: Levels node is drawn without a color
When multiple levels node exist, only one is drawn correctly.

The levels node has the class 'Output' and therefore is shown dark red.
Although the color is not too critical, the color drawing is based on
whether the output node is active or not.

To fix this we change the class to Converter (same as math node), so the
node header is now blue, but always drawn correctly.