UI: Better error message when entering modes on linked/packed data
In 5.0, assets from the asset browser are impored using Packing by
default. This means entering a mode (edit mode, sculpt mode, etc.) would
show an error. For such a basic use case the error message should be
more useful than it was. It should explain why entering the mode isn't
possible, and how to address that.