Projects: Dynamic navigation sections in settings from panel categories
Instead of hard-coding the available navigation items in the Project
settings, allow defining sections using the `bl_category` of a panel.
This is how tabs in sidebars are collected dynamically as well.

This makes it easy to not just add custom panels, but also custom
navigation items. This can be useful for studios to add own project
setting UIs in a well integrated manner.

However, this is also just a generally useful way of extending the UI
from Add-ons. Plus, this makes it a lot easier to add new navigation
items for core developers too.

Further:
- Separate between supporting categories for navigation, and drawing the
  categories as tabs.
- Refactor how it's defined which parts of the UI can use categories
  (and category tabs). Was a bit hacky before, now it's defined with a
  simple region type flag.