Release note tools: Factor out the gathering and printing of release notes information
This is a small refactor of part of the release notes tool so the
gathering of information that should be printed to terminal, and the
actual printing of the informaiton is split into two seperate steps.

This allows the "gathering" function to be reused by other Python
scripts. For example, the script that presents the output of the
bug fixes per release script in a friendly mannor for triagers at
https://projects.blender.org/Alaska/needs-triager-action/issues/2
https://projects.blender.org/Alaska/needs-triager-action/issues/3

This change should have no functional impact on the bug fixes per
release script, outside of the occasional white space change in the
output of the script.