Refactor: Compositor: Use callbacks for writing outputs
This patch refactors how outputs are eventually written by the context.
Instead of getting a result from the context, output operations now
provide the final result and the context can copy it to the needed data.

This should improves performance in some cases due to avoidance of
double copies. It also makes it easier to add new contexts with
different output requirements.