Fix #149494: Performance regression with "orco tangents" and EEVEE
Caused by 9a672cbe52638240e30a0f78d3618b00cc093502

There is specific code to invalidate tangents and UV map vertex buffers
because we pack multiple attributes into the same vertex buffer, and we
don't necessarily want to rebuild the VBOs whenever the set of requested
UV maps/tangents changes at all. "Orco tangents" or non-UV-map tangents
are also packed into the same buffer. When you switch to EEVEE and then
back to workbench, orco tangents are requested but then not needed
anymore, causing `cd_used` and `cd_needed` not to match anymore.

Currently this causes the tangents to be rebuilt and re-uploaded on
on every redraw, but previously the code would skip this invalidation
because it skipped invalidating the VBOs when `cd_used` contained
something that `cd_needed` didn't.

This sounds very complicated (and it is, the system shouldn't be so
weird), but ultimately it's just a small logic error in the line that
used to read `if (cd_overlap == false || attr_overlap == false) {`.

Also fix an issue in the `DRW_MeshCDMask` equality operator,
though I don't know of that causing visible issues so far.