Fix #145073: Strip retiming is imprecise
Caused by dadbb5c5f, which moved the last key to the start of the last
frame of the strip content. However this caused, that the last frame
can not be retimed.
From user perspective, the bug was, that if you set 10 frame strip speed
to 50% it would be only 19 frame long,

Note, that #145073 is multi report so there should be another fix for
the issue mentioned there.