Subdiv: Avoid function call overhead for face propagation
Previously there was a function call for every subdivided faces, then a
function call for each of the attributes. Face attribute propagation is
actually very simple, and we can use existing utility functions to
duplicate values for the subdivided faces.

I tested this in a fairly extreme case with 100 face attributes on a 60
thousand face mesh and observed a 1.2x performance increase in overall
CPU subdivision performance (from 134 to 112 ms).

I'll note that this slightly breaks the `foreach_subdiv_geometry`
abstraction in that it's not just calling a callback per subdivided
element anymore. I think that's generally a good direction to take
the code though, to structure it in a more data-oriented way.