Cleanup: VSE: Standardize `nullptr` scene, ed checks
In #145145 `editing_get` was updated to be able to deal with `nullptr`
scenes, but this effectively made it possible to get away with only
checking for a valid `Editing`. This was mentioned briefly in discussion
on #145968, where it was agreed that `nullptr` check responsibility
should be clearer.

With VSE's current design, either of the following is possible:
- No sequencer scene, no Editing. E.g. newly created Video Editing
  workspace.
- Sequencer scene, but no Editing. This can happen if creating a new
  blank scene with the ID selector at the window level, then choosing
  that scene to be the sequencer scene.

So it makes sense to have to check for both. Letting callers get away
with just checking for the latter can be a bit confusing, and in
practice, almost all code already does both checks, so this patch just
standardizes the behavior.