WIP: Fix: Crashes in modal operators when scene context changes
This is an attempt at fixing crashes in modal operators that work
on the assumption that the scene context doesn't change between
modal runs.

Ideally these operators would handle this properly since it is okay
for the context to change, but currently this is a general issue across
multiple modal operators it seems like.

This introduces a `Scene` in the `wmEventHandler_Op` context which is
stored when the event handler is created. Then this scene is set
in the context for every modal execution (if the scene is different from
the active scene in the window).

Currently, this is a failed attempt because it breaks the `frame_change`
operator (frame scrubbing) when using the `Sync Scene Time` feature.
The frame change operator can change the scene inside the modal loop,
which is immediatly restored by the operator code after the modal
function executes.